openid_connect: Move code around
We now have all the openid providers at the same place
