+++ config/settings/base.py
@@ -77,8 +77,8 @@ LOCAL_APPS = [
     "itou.job_applications",
     "itou.approvals",
     "itou.eligibility",
-    "itou.france_connect",
-    "itou.inclusion_connect",
+    "itou.openid_connect.france_connect",
+    "itou.openid_connect.inclusion_connect",
     "itou.invitations",
     "itou.external_data",
     "itou.metabase",
