+++ config/urls.py
@@ -47,9 +47,9 @@ urlpatterns = [
     # PEAMU URLs.
     path("accounts/", include("itou.allauth_adapters.peamu.urls")),
     # FranceConnect URLs.
-    path("franceconnect/", include("itou.france_connect.urls")),
+    path("franceconnect/", include("itou.openid_connect.france_connect.urls")),
     # Inclusion Connect URLs.
-    path("inclusion_connect/", include("itou.inclusion_connect.urls")),
+    path("inclusion_connect/", include("itou.openid_connect.inclusion_connect.urls")),
     # --------------------------------------------------------------------------------------
     # API.
     path("api/v1/", include("itou.api.urls", namespace="v1")),
