+++ itou/openid_connect/france_connect/apps.py
@@ -2,5 +2,5 @@ from django.apps import AppConfig
 
 
 class FranceConnectConfig(AppConfig):
-    name = "itou.france_connect"
+    name = "itou.openid_connect.france_connect"
     verbose_name = "France Connect"
