+++ itou/openid_connect/inclusion_connect/apps.py
@@ -2,5 +2,5 @@ from django.apps import AppConfig
 
 
 class InclusionConnectConfig(AppConfig):
-    name = "itou.inclusion_connect"
+    name = "itou.openid_connect.inclusion_connect"
     verbose_name = "Inclusion Connect"
