+++ itou/openid_connect/inclusion_connect/constants.py
@@ -1,5 +1,3 @@
-import datetime
-
 from django.conf import settings
