+++ itou/openid_connect/inclusion_connect/models.py
@@ -1,7 +1,8 @@
 import dataclasses
 
-from itou.common_apps.openid_connect.models import OIDConnectState, OIDConnectUserData
 from itou.users.enums import IdentityProvider
+
+from ..models import OIDConnectState, OIDConnectUserData
 
 
 class InclusionConnectState(OIDConnectState):
